requirements.txt
*Install  All these Libraries *
tensorflow==2.11.0
numpy
p-utils
Pillow
jupyter
keras
pandas
matplotlib
seaborn
scikit-learn
opencv-python
flask